+++ xiao3sh/src/main.rs
@@ -1,45 +1,11 @@
 use std::collections::HashMap;
-use std::env;
 use std::io::{self, Error, ErrorKind, Write};
-use std::path::Path;
 use std::process::Command;
+
+mod builtins;
 
 fn split_line(line: &str) -> impl Iterator<Item = &str> {
     line.trim().split_whitespace()
-}
-
-fn builtin_exit(args: &[String]) -> io::Result<()> {
-    if args.len() > 1 {
-        return Err(Error::new(ErrorKind::InvalidInput, "too many arguments"));
-    }
-    let code = args.get(0).map_or(0, |n| n.parse::<i32>().unwrap_or(0));
-    std::process::exit(code)
-}
-
-fn builtin_cd(args: &[String]) -> io::Result<()> {
-    if args.len() > 1 {
-        return Err(Error::new(ErrorKind::InvalidInput, "too many arguments"));
-    }
-    if args.len() < 1 {
-        // TODO implement to move home directory after supporting environment variables
-        return Ok(());
-    }
-    let path = Path::new(args.get(0).unwrap());
-    match env::set_current_dir(path) {
-        Ok(()) => Ok(()),
-        Err(err) => {
-            // FIXME output more kind message
-            eprintln!("failed to change directory: {} ({:?})", path.display(), err);
-            Ok(())
-        }
-    }
-}
-
-fn create_builtins() -> HashMap<String, fn(&[String]) -> io::Result<()>> {
-    let mut r: HashMap<String, fn(&[String]) -> io::Result<()>> = HashMap::new();
-    r.insert("exit".to_string(), builtin_exit);
-    r.insert("cd".to_string(), builtin_cd);
-    r
 }
 
 fn launch(cmd: &str, args: &[String]) -> io::Result<()> {
@@ -80,7 +46,7 @@ where
 
 fn main_loop() -> io::Result<()> {
     let mut input = String::new();
-    let builtins = create_builtins();
+    let builtins = builtins::create_builtins();
     loop {
         print!("> ");
         io::stdout().flush().unwrap();
